remove stack too deep

+++ contracts/strategies/baseStrategy/v1/interfaces/IBSSwapHelperFacet.sol
@@ -5,5 +5,5 @@ pragma solidity ^0.8.18;
 interface IBSSwapHelperFacet {
     function quote(address src, address dst, uint256 amount) external;
     
-    function swap(address src, address dst, uint256 amount, uint256 slippage) external;
+    function swap(address src, address dst, uint256 amount, uint8 slippage) external;
 }
